Documentacio y mas fixes del codigo
Documentacio y mas fixes del codigo

+++ Scripts_Docu/Store Procedures/SQLQuery_Comprobante_Pago_SP.sql
@@ -4,27 +4,56 @@ CREATE PROCEDURE SPI_Comprobante_Pago
 @total int,
 @idPropiedad int
 AS 
-BEGIN
+BEGIN TRY
+	DECLARE @retValue int;
 	IF @idPropiedad is null
 		BEGIN
-		RAISERROR('Propiedad no encontrada',10,1)
-		RETURN -10
+			RAISERROR('Propiedad no encontrada',10,1)
+			SET  @retValue =  -10
 		END
 	ELSE 
 		BEGIN
-		INSERT INTO dbo.Comprobante_Pago(fecha, total, idPropiedad) 
-		VALUES(@fecha, @total, @idPropiedad)
-		RETURN SCOPE_IDENTITY()
+			INSERT INTO dbo.Comprobante_Pago(fecha, total, idPropiedad) 
+			VALUES(@fecha, @total, @idPropiedad)
+			RETURN SET  @retValue = SCOPE_IDENTITY()
 		END
-END
+	RETURN @retValue
+END TRY
+BEGIN CATCH
+	DECLARE 
+		@Message varchar(MAX) = ERROR_MESSAGE(),
+		@Severity int = ERROR_SEVERITY(),
+		@State smallint = ERROR_STATE()
+ 
+	RAISERROR( @Message, @Severity, @State) 
+END CATCH
 
 --Delete
 CREATE PROCEDURE SPD_Comprobante_Pago
 @idPropiedad int
 AS
-BEGIN
-	DELETE FROM dbo.Comprobante_Pago WHERE idPropiedad = @idPropiedad
-END
+BEGIN TRY
+	DECLARE @retValue int;
+	IF EXISTS (SELECT * FROM dbo.Comprobante_Pago WHERE idPropiedad = @idPropiedad)
+		BEGIN
+			SET @retValue =  (SELECT id FROM dbo.Comprobante_Pago WHERE idPropiedad = @idPropiedad);
+			DELETE FROM dbo.Comprobante_Pago WHERE idPropiedad = @idPropiedad
+		END
+	ELSE
+		BEGIN
+			RAISERROR('Numero de finca ya registrado', 10, 1)
+			SET @retValue =  -14;
+		END
+	RETURN  @retValue;
+END TRY
+BEGIN CATCH
+	DECLARE 
+		@Message varchar(MAX) = ERROR_MESSAGE(),
+		@Severity int = ERROR_SEVERITY(),
+		@State smallint = ERROR_STATE()
+ 
+	RAISERROR( @Message, @Severity, @State) 
+END CATCH
 
 --Prueba
 DECLARE @id int
